:recycle: refactor: separate the video repository into stores
